Implements multi-valued items in make-table.rb
Some headers appear in multiple rows on the tables, with different
values for each. This implements organizing those values by
"rank" where possible and writing them to a table multi-headers.tsv

Note that in practice, there is a large amount of "rank collision" messages
output when this is run as it turns out there's a lot of ranked values
which have the same rank.